torchtitan/requirements.txt
torchdata >= 0.8.0
datasets >= 2.21.0
tomli >= 1.1.0 ; python_version < "3.11"
tensorboard
tiktoken
blobfile
tabulate
wandb
expecttest==0.1.6
pytest==7.3.2
pytest-cov
pre-commit
tomli-w >= 1.1.0
